[UPDATE]:update page.jsx of restaurants and add restaurant.jsx
